chore: remove dead morphometrics_config and seg_to_mesh modules
ConfigEditor and SegToMeshSubmissionWidget are not wired into main.py;
ConfigYAMLPreserver is unused by any live code path. The only live
exports from morphometrics_config were IntraListEditor and
InterDictEditor, used by distance_tab.

Inline both editors directly into distance_tab. While inlining,
InterDictEditor now keys self.entries by a monotonic token instead
of the user-typed membrane name — the previous version collided on
the empty key ('') for every button-driven Add, silently overwriting
prior entries in get_values() and routing Remove clicks to the wrong
row.
